+++ src/app/features/restaurant/pages/rooms/rooms.ts
@@ -28,8 +28,8 @@ import { omitKeys } from '@/lib/helpers';
   templateUrl: './rooms.html',
   styleUrl: './rooms.css',
 })
-export class Rooms extends BaseComponent<IRoomRowResponse> { 
-   currentItem: IRoomDtoResponse | null = null;
+export class Rooms extends BaseComponent<IRoomRowResponse> {
+  currentItem: IRoomDtoResponse | null = null;
 
   initialFormValue = {
     id: this.fb.control<number>(0, []),
@@ -106,10 +106,32 @@ export class Rooms extends BaseComponent<IRoomRowResponse> {
     });
   }
 
-  deleteRoom(id: number) {
-    this.roomService.delete(id).subscribe({
-      next: ()=>{
-        this.resetState();
+  deleteRoom(id: number, event: Event) {
+    this.confirmationService.confirm({
+      target: event.target as EventTarget,
+      message: 'هل انت متاكد من حذف الغرفة؟',
+      header: 'حذف الغرفة',
+      icon: 'pi pi-info-circle',
+      rejectLabel: 'الغاء',
+      rejectButtonProps: {
+        label: 'الغاء',
+        severity: 'secondary',
+        outlined: true,
+      },
+      acceptButtonProps: {
+        label: 'حذف',
+        severity: 'danger',
+      },
+
+      accept: () => {
+        this.roomService.delete(id).subscribe({
+          next: () => {
+            this.resetState();
+          },
+        });
+      },
+      reject: () => {
+        this.messageService.add({ severity: 'error', summary: 'الغاء', detail: 'لقد قمت بالغاء الحذف' });
       },
     });
   }
